ConvertToFloat now calculates INT instead of FLOAT, improved performance

+++ Src/ConvertToFloat.h
@@ -24,6 +24,6 @@ private:
 	void convYV24ToFloat(const byte *py, const byte *pu, const byte *pv,
 		unsigned char *dst, int pitch1Y, int pitch1UV, int pitch2, int width, int height, IScriptEnvironment* env);
 	void convRgbToFloat(const byte *src, unsigned char *dst, int srcPitch, int dstPitch, int width, int height, IScriptEnvironment* env);
-	void convFloat(int y, int u, int v, unsigned char *out);
+	void convFloat(byte y, byte u, byte v, unsigned char *out);
 	VideoInfo viRGB;
 };
